Migrate messages UI to inferno
